Build for 10 sec Person Spawning
This commit is NOT intended to be merged into the base game. It is purely for testing persons, and as such, has person ships spawn every 10secs instead of the usual 10min.

Also, note that I let mingw install in its default location, which means my folder path is not the same as the one setup in the default .cbp file. This branch has my own file locations.

Other than those two changes, this branch will be kept up to date with continuous, and won't have other changes.