Allow some one-time duplicates of non-conversational missions
We most definitely want to avoid reoffering a mission with a conversation, but these two aren't those such missions, so it's less important that they be blocked from offering if the older version of the mission offered. This cuts down on the amount of compatibility code that needs to be maintained.