To facilitate testing purposes, display mass and fuel per HD jump
Currently a temporary measure to facilitate testing.

That being said, I'm not averse to putting conditions on these so players can choose to make them visible, and then having them as a permanent feature. Useful for contributors, and I could see some players being interested in seeing this info live.