Scattering the HUD display outfits around a bit
The two hyperjump ones and the mass into Deep and Remnant.
The two scramdrive ones into everywhere that sells scram drives.
THe two jump drive ones into the Remnant Salvage one (so isn't visible immediately, but will be available after the Player has ample access to Korath tech, including JD.
The Fixed amount fuel bar is available in every "common" outfit outfitter, which is found all over the place, and includes most places that the HD is sold (although not quite all of them). It should be by far the easiest display to get, which makes sense, since it is the simplest.