Adding Lateral Ionic Thrusters
The X1050 gets built-in lateral thrust equal to 90% of its forward thrust, being an all-in-one engine.

Note on Lateral Thrusters: My initial values here are to have each set of lateral thrusters be based on the thruster of the same "rank".  So the tiny lateral has values equivalent to about 90% of a tiny thruster. The small lateral thruster is roughly 80% of a small thruster, etc. This continues by 10% increments down to the Huge lateral thrusters that are equivalent of 50% of the huge thrusters. This percentage is applied equally to mass, outfit space, engine capacity, lateral thrust, lateral thrusting energy, and lateral thrusting heat.

I'm not set on the values, but they seemed a good place to start. Note that once these are installed on all stock ships, those ships will have noticeably better lateral thrust than they used to under the ratio system. Only marginally at the fast end, but quite significantly at the heavy end.