Remnant Lateral Thrusters
This breaks somewhat with the previous pattern, as I used a 20% drop to account for the lack of the reverse thrust capability. So the Remnant use the 0.9, 0.8, and 0.7 pattern for thrust and costs, with a 0.7, 0.6, and 0.5 for the mass and space. Maybe this should be more, I'm not sure. The Remnant thruster ecosystem needs some work to make it more interesting. Oh, and there's no "huge" variant. Which would be nice to have.