Marauders get laterals now.
I'll be honest, the amount of extra outfit and engine space that some of these ships get seems... A bit extreme. I haven't actually compared with the original ships, but if feels very high. I suspect that the engine variants in particular will be able to reach ludicrous speeds. Worth keeping an eye on, anyway.