More adjustments to ensure that all ships have lateral thrust
The only exceptions are derelict ships and the science drone.

Until such time as the AI is significantly improved, this should be the general rule: All ships have lateral thrust, period. Minimum 50% for slow, heavy, and huge ships, 90% for agile, light, and tiny ships. The minimum is pretty close to a hard rule, the maximum is just a guideline. A ship certainly could have equal or better lateral thrust than forward thrust. It should be rare, but could be the case.