Thrust Reduction and working on getting the attributes to display.
Thrust Reduction works. that part is good, and its results display properly in Player Info, too. And the ship actually handles as if it is having the specified impact.

But it doesn't display properly ( x60) in the outfit info, and the info on the selected ship in the outfitter also doesn't take it into account.

The lateral Ratio *also* displays as x60 its specified value in the oufitter.

(0.5 displays at 30)