Progress!
Vitalchip had a moment of inspiration that solved why the map kept getting erased, and now the atribute changes are making there way all the way through the code, and at least in the case of hull, temporarily changing the hull value that is stored in the player's save file but outside the "attributes" section of their ship.

So it's not entirely solved yet, but massive progress.

Co-Authored-By: vitalchip <65342577+vitalchip@users.noreply.github.com>