Removing scaling from Marauder arrow and bounder
If these were to be scaled, then the arrow and bounder would need to be scaled too, and that might lead to other smaller ships being scaled... And I don't want to go down that rabbit hole right now. So the scaling on them is being removed. I'm not averse to them being revised in a later PR, just not in this one.