Giving James a more meaningful gift
Gives the "James' Engine Mod" at the end of the mission string, which gives 10% acceleration and turn multiplication.

Not sure if those work as +10%, or 10%, though. As in, it is set to 0.1. Is that going to be setting the player's acceleration to 10% of what it was, or to 110%? I'm thinking probably the former...