Just adding slots to get it to be "correct"
I don't fully understand everything about how the Avgi work yet, but looks complex and realistic. I'm liking it, but it's going to take some work. On the other hand, kind of tempted to just leave as is since it looks like it's got the most hard-science work put into its engines out of everything we've got in-game.