Added outfits from sales for Avgi, Bunrodea, Coalition, Gegno, Hai, Incipias, Rulei, Sheragi, Successors, Ka'het, Korath
Pug and Wanderers are deliberately omitted for story purposes. Ka'het and Korath are on separate lists for story reasons. If there are any outfits that are not on sales list, they didn't get included, but they should be. I just don't have time to manually go through every outfit file on every faction.