Multimodal Armor special handling
Adds a cost to the multimodal armor, and "all non-remnant" price reduction that zeroes it out.

Multimodal armor is NOT actually an outfit, it is a mode the armor can be set to. The cost is the price in materials, energy, labor, etc. for the outfitter to transform the armor into that mode. Resetting it back to base is free and can be done anywhere. The Remnant aren't paying for an outfit, in this case, they are paying to get to watch the transformation happen.